chain: rosetta: extract transaction-related code to new module
`mod.rs` is around 1500 lines long.  It’s time to split it especially
as code for constructing transaction objects is about to become longer
and more complex.

Issue: https://github.com/near/nearcore/issues/5173
